Move all inline parameter handling into a dedicated module
